pyproject.toml
[project]
name = "mednot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "passlib>=1.7.4",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
